split bytecode in mod files
